Split components and common code in directories
