+++ crosswrd-app/src/common/Lights.tsx
