+++ crosswrd-app/src/common/Spiral.tsx
