+++ crosswrd-app/src/index.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { render } from "react-dom";
 import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import App from "./App";
+import App from "./components/App";
 import reportWebVitals from "./reportWebVitals";
 
 render(
